[Segment Resource] UseStateForUnknown modifier on id
While testing internally, I found that the provider wants to recreate a
ruleset if the corresponding segment is updated at all. This is not
necessary and is actively harmful. It would mean that the rules are
unapplied temporarily (even if only for a few seconds).

The root cause is that terraform doesn't know the `id` is constant once
created. This PR uses a modifier from the framework to specify that.